Leave group chat

+++ src/components/DeleteConnection.jsx
@@ -2,18 +2,22 @@ import React from 'react'
 import apiClient from '../services/ApiClient'
 import { useNavigate } from 'react-router-dom';
 
-const DeleteConnection = ({ connectionId, setDeleteConnectionSelected }) => {
+const DeleteConnection = ({ connectionId, chatId, setDeleteConnectionSelected }) => {
     const navigate = useNavigate();
 
-    const deleteConnection = async () => {
-        await apiClient.deleteConnection(connectionId)
+    const handler = async () => {
+        if (connectionId) {
+            await apiClient.deleteConnection(connectionId)
+        } else if (chatId) {
+            await apiClient.leaveGroupChat(chatId)
+        }
         navigate("/");
     }
 
     return (
         <div className="delete-connection">
             <button className="btn btn-sm btn-primary" onClick={() => setDeleteConnectionSelected(false)}>Cancel</button>
-            <button className="btn btn-sm btn-danger" onClick={deleteConnection}>Confirm delete</button>
+            <button className="btn btn-sm btn-danger" onClick={handler}>Confirm</button>
         </div>
     )
 }
